requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
boto3==1.42.13
botocore==1.42.13
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
Nuitka==2.8.9
ordered-set==4.1.0
pillow==12.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-magic==0.4.27
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
s3transfer==0.16.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==14.1.0
Twisted==25.5.0
txaio==25.9.2
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
zope.interface==8.1.1
zstandard==0.25.0
